dashboard/requirements.txt
Flask==3.0.0
openai>=1.12.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
Flask-Dance==7.0.0
requests==2.31.0
boto3>=1.34.0
psycopg2-binary>=2.9.9

